Move RevenueStat to Analytics.kt file
