--- WooCommerce/src/main/kotlin/com/woocommerce/android/model/AnalyticStat.kt
@@ -1,11 +1,0 @@
-package com.woocommerce.android.model
-
-sealed class AnalyticStat {
-    data class RevenueStat(
-        val totalValue: Double,
-        val totalDelta: Int,
-        val netValue: Double,
-        val netDelta: Int,
-        val currencyCode: String?
-    ) : AnalyticStat()
-}
