+++ WooCommerce/src/main/kotlin/com/woocommerce/android/model/Analytics.kt
@@ -1,0 +1,9 @@
+package com.woocommerce.android.model
+
+data class RevenueStat(
+    val totalValue: Double,
+    val totalDelta: Int,
+    val netValue: Double,
+    val netDelta: Int,
+    val currencyCode: String?
+)
